requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
streamlit
requests
pytest
matplotlib
